requirements.txt
fastapi==0.110.3
uvicorn==0.30.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
pyyaml==6.0.1
mlflow==2.13.1
joblib==1.4.2
pydantic==2.7.1

